namespace (Patchwork\) and convert to PSR-4 (fixes #11)
Previously JSqueeze claimed PSR-0 compliance in composer.json
but was not in fact compliant, as PSR-0 requires at least one
level of namespacing, bare classes are not PSR-0 compliant.
As PSR-4 has now deprecated PSR-0, convert to PSR-4 while
we're about it, which has the handy bonus of meaning the class
file does not need to be moved at all.

This is an obvious BC break: all consumers have to switch from
calling 'new \JSqueeze()' to 'new \Patchwork\JSqueeze()'.