Fix for empty statement after `else` at end of block
Remove any `else` which ends up with nothing following it at the end of
a block, e.g. in
```
function f(x) {
if (x)
return x;
else
;
}
```
Previously, the semicolon would be removed but not the `else` keyword,
resulting in invalid code.  (This was only an issue at the end of a code
block, i.e. immediately before a `}`.)