Fix `;` incorrectly inserted after for/if/while and other keywords followed by prefix operator
The previous change adds semicolons that shouldn't be in constructs like
`while(x)\n--x`, `else\n--x`, `do\n--x` and `return --\nx`.

This is addressed for the `for`/`if`/`while` case by marking a potential
semicolon after a closing bracket with a `#` symbol, and handling it
later when removing semicolons in the loop where the state of preceding
`for`/`if`/`while` expressions is maintained.  To cater for `do..while`
statements, additional state is now maintained so as to correctly
determine whether a `while` corresponds to a `do`.  Additionally, this
fixes a pre-existing bug where a semicolon was not always added after a
`do..while`.

For other keywords, their presence can simply be checked with a regex so
that e.g. `else++` is not treated as postincrement and `else\n++x` not
as separate statements.

Also spotted that the `substr` length wasn't enough for excluding
`continue` at the end of a variable name when adding a semicolon after,
and corrected this.