Optimize final `str_replace` restoring strings
For a large source file (1Mb+) with a large number of strings (10,000+),
JSqueeze can take a long time to complete due to the inefficiency of
`str_replace` with multiple replacements (restoring the original JavaScript
strings and regular expressions in the code).

As the code only needs to be processed once to make all the string replacements,
`preg_replace` is much more efficient here.  In the case of a 5Mb source file
with 20,000 strings it is approx. 100x faster for the string replacements,
making squeezing about 8x faster overall.

Even for small files, this appears to be a performance improvement: the unit
tests seem to be taking about 5% less time.